% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_to_sqlite.R
\name{access_to_sqlite}
\alias{access_to_sqlite}
\title{moves data from fieldmap database to SQLite database}
\usage{
access_to_sqlite(source_db, new_db)
}
\arguments{
\item{source_db}{name of fieldmap/access database (with specific fieldmap structure) including path}

\item{new_db}{name and path of SQLite database to be made}
}
\value{
A SQLite database with the specified name is saved at the specified path
}
\description{
This function moves all relevant tables from an unindexed Access database to a SQLite database with indexes, for performance reasons.
}
