% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deadw_decay_plot.R
\name{calc_deadw_decay_plot}
\alias{calc_deadw_decay_plot}
\title{aggregate parameters by decay stage, plot and year}
\usage{
calc_deadw_decay_plot(plotinfo, data_deadwood = NA, data_dendro_calc = NA)
}
\arguments{
\item{plotinfo}{dataframe on surveyed plots with variables \code{plot_id},
\code{plottype}, \code{forest_reserve}, \code{survey_trees}, \code{survey_deadw}, \code{period} and
\code{year_dendro}
(output of function \code{load_plotinfo()})}

\item{data_deadwood}{dataframe on logs with variables \code{plot_id}, \code{plottype},
\code{date_dendro}, \code{species}, \code{decaystage}, \code{calc_volume_m3}, \code{period} and \code{year}
(output of function \code{load_data_deadwood()})}

\item{data_dendro_calc}{dataframe on stems (shoots and trees) as given from
the function \code{calc_variables_stem_level()}}
}
\value{
dataframe with columns \code{plot}, \code{year}, \code{decaystage}, \code{vol_log_m3_ha}
}
\description{
This function calculates for each plot and year the volume logs and standing
dead wood per hectare and per decay stage.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")

data_deadwood <- load_data_deadwood(path_to_fieldmapdb)
data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models()
data_stems_calc <- calc_variables_stem_level(data_stems, height_model)
data_dendro_calc <- calc_variables_tree_level(data_dendro, data_stems_calc)
plotinfo <- load_plotinfo(path_to_fieldmapdb)
calc_deadw_decay_plot(plotinfo, data_deadwood, data_dendro_calc)

}
