% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dendro_plot.R
\name{calc_dendro_plot}
\alias{calc_dendro_plot}
\title{aggregate parameters by plot and year}
\usage{
calc_dendro_plot(data_dendro_calc, data_deadwood, plotinfo)
}
\arguments{
\item{data_dendro_calc}{dataframe on tree measures with variables \code{plot_id},
\code{plottype}, \code{tree_measure_id}, \code{date_dendro}, \code{dbh_mm}, \code{height_m},
\code{species}, \code{alive_dead}, \code{decaystage}, \code{vol_tot_m3}, \code{basal_area_m2},
\code{period}, \code{old_id}, \code{year}, \code{subcircle}, \code{plotarea_ha},...
(output of function \code{calc_variables_tree_level()})}

\item{data_deadwood}{dataframe on logs with variables \code{plot_id}, \code{plottype},
\code{date_dendro}, \code{species}, \code{decaystage}, \code{calc_volume_m3}, \code{period} and \code{year}
(output of function \code{load_data_deadwood()})}

\item{plotinfo}{dataframe on surveyed plots with variables \code{plot_id},
\code{plottype}, \code{forest_reserve}, \code{survey_trees}, \code{survey_deadw}, \code{period} and
\code{year_dendro}
(output of function \code{load_plotinfo()})}
}
\value{
dataframe with columns plot, year, number_of_tree_species,
number_of_trees_ha, basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot and year some values per hectare:
number of tree species, number of trees, basal area and volume.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")

data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_deadwood <- load_data_deadwood(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models()
data_stems_calc <- calc_variables_stem_level(data_stems, height_model)
data_dendro_calc <- calc_variables_tree_level(data_dendro, data_stems_calc)
plotinfo <- load_plotinfo(path_to_fieldmapdb)
calc_dendro_plot(data_dendro_calc, data_deadwood, plotinfo)

}
