% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_regeneration_core_area_height_species.R
\name{calc_reg_core_area_height_spec}
\alias{calc_reg_core_area_height_spec}
\title{aggregate regeneration parameters by plot, height, species and year}
\usage{
calc_reg_core_area_height_spec(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
\code{plot_id}, \code{plottype}, \code{subplot_id}, \code{height_class}, \code{species},
\code{nr_of_regeneration}, \code{rubbing_damage_number}, \code{period}, \code{year}, \code{subcircle},
\code{plotarea_ha}, \code{min_number_of_regeneration} and \code{max_number_of_regeneration}.}
}
\value{
dataframe with columns \code{plot}, \code{species}, \code{year}, \code{height},
\code{nr_of_subplots_with_regeneration}, \code{perc_subplots_with_regeneration},
\code{approx_rubbing_damage_perc}, \code{mean_number_of_regeneration_ha},
\code{lci_number_of_regeneration_ha}, \code{uci_number_of_regeneration_ha} and
\code{approx_nr_regeneration_ha.}
}
\description{
This function calculates for each plot, height, species and year the number
of regeneration per ha (or interval with mean and confidence interval
using a log transformation), the number and percentage of
subplots in which the species is regenerating and
the approximate rubbing damage percentage per hectare.
This calculation is designed for core areas, that consist of different
subplots.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration_CA <-
  load_data_regeneration(path_to_fieldmapdb, plottype = "CA")
calc_reg_core_area_height_spec(data_regeneration_CA)

}
