% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_reg_core_area_species.R
\name{calc_reg_core_area_species}
\alias{calc_reg_core_area_species}
\title{aggregate regeneration parameters by plot, species and year}
\usage{
calc_reg_core_area_species(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
\code{plot_id}, \code{plottype}, \code{subplot_id}, \code{height_class}, \code{species},
\code{nr_of_regeneration}, \code{rubbing_damage_number}, \code{period}, \code{year}, \code{subcircle},
\code{plotarea_ha}, \code{min_number_of_regeneration} and \code{max_number_of_regeneration}.}
}
\value{
dataframe with columns \code{plot}, \code{species}, \code{year},
\code{nr_of_subplots_with_regeneration}, \code{perc_subplots_with_regeneration},
\code{mean_number_established_ha},
\code{lci_number_established_ha}, \code{uci_number_established_ha},
\code{mean_number_seedlings_ha},
\code{lci_number_seedlings_ha}, \code{uci_number_seedlings_ha},
\code{mean_rubbing_damage_perc_established},
\code{lci_rubbing_damage_perc_established},
\code{uci_rubbing_damage_perc_established},
\code{mean_rubbing_damage_perc_seedlings}, \code{lci_rubbing_damage_perc_seedlings},
\code{uci_rubbing_damage_perc_seedlings},
\code{approx_nr_established_ha}, \code{approx_nr_seedlings_ha},
\code{approx_rubbing_damage_perc_established},
\code{approx_rubbing_damage_perc_seedlings}.
}
\description{
This function calculates for each plot, species and year the number of
seedlings and established regeneration per ha (or interval with mean
and confidence interval using a log transformation), the number and
percentage of subplots in which the species is regenerating and
the approximate rubbing damage percentage per hectare.
This calculation is designed for core areas,that consist of different
subplots.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration_CA <-
  load_data_regeneration(path_to_fieldmapdb, plottype = "CA")
calc_reg_core_area_species(data_regeneration_CA)

}
