% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_reg_plot.R
\name{calc_reg_plot}
\alias{calc_reg_plot}
\title{calculate species number by plot and year}
\usage{
calc_reg_plot(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
\code{plot_id}, \code{plottype}, \code{subplot_id}, \code{height_class}, \code{species},
\code{nr_of_regeneration}, \code{rubbing_damage_number}, \code{period}, \code{year}, \code{subcircle},
\code{plotarea_ha}, \code{min_number_of_regeneration} and \code{max_number_of_regeneration}.}
}
\value{
dataframe with columns \code{plot}, \code{subplot},
\code{year}, \code{date_regeneration}, \code{period},
\code{number_of_tree_species}, \code{nr_of_tree_species_established},
\code{mean_number_established_ha}, \code{lci_number_established_ha},
\code{uci_number_established_ha},
\code{mean_number_seedlings_ha}, \code{lci_number_seedlings_ha},
\code{uci_number_seedlings_ha},
\code{mean_rubbing_damage_perc_established},
\code{lci_rubbing_damage_perc_established},
\code{uci_rubbing_damage_perc_established},
\code{mean_rubbing_damage_perc_seedlings}, \code{lci_rubbing_damage_perc_seedlings},
\code{uci_rubbing_damage_perc_seedlings},
\code{approx_nr_established_ha}, \code{approx_nr_seedlings_ha},
\code{approx_rubbing_damage_perc_established},
\code{approx_rubbing_damage_perc_seedlings}.
}
\description{
This function calculates for each plot and year the number of species, total
number of seedlings and established regeneration (or interval with mean
and confidence interval using a log transformation) and
approximate rubbing damage percentage for seedlings and established
regeneration.
For core area plots, these variables are calculated for each subplot.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration <- load_data_regeneration(path_to_fieldmapdb)
calc_reg_plot(data_regeneration)

}
