% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot_height.R
\name{calc_reg_plot_height}
\alias{calc_reg_plot_height}
\title{calculate species number by plot, tree height class and year}
\usage{
calc_reg_plot_height(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
\code{plot_id}, \code{plottype}, \code{subplot_id}, \code{height_class}, \code{species},
\code{nr_of_regeneration}, \code{rubbing_damage_number}, \code{period}, \code{year}, \code{subcircle},
\code{plotarea_ha}, \code{min_number_of_regeneration} and \code{max_number_of_regeneration}.}
}
\value{
dataframe with columns \code{plot}, \code{subplot}, \code{year}, \code{period},
\code{height_class}, \code{number_of_tree_species}, \code{approx_rubbing_damage_perc},
\code{mean_number_of_regeneration_ha},
\code{lci_number_of_regeneration_ha}, \code{uci_number_of_regeneration_ha} and
\code{approx_nr_regeneration_ha}.
}
\description{
This function calculates for each plot, tree height class and year
the number of species, total number of regeneration (or interval with mean
and confidence interval using a log transformation) and
approximate rubbing damage percentage for regeneration.
For core area plots, these variables are calculated for each subplot.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration <- load_data_regeneration(path_to_fieldmapdb)
calc_reg_plot_height(data_regeneration)

}
