% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stem_volume.R
\name{calc_stem_volume}
\alias{calc_stem_volume}
\title{Calculate stem volume including bole and crown volume}
\usage{
calc_stem_volume(data_stems)
}
\arguments{
\item{data_stems}{dataframe on stems (shoots and trees) in which the height
is already corrected using the height models (if necessary)}
}
\value{
Dataframe with ...
}
\description{
This internal helper function calculates the bole and crown volume to be used
in functions `calc_variables_stem_level()` and `calc_intact_deadwood()`.
}
