% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_variables_stem_level.R
\name{calc_variables_stem_level}
\alias{calc_variables_stem_level}
\title{Calculate additional variables on stem level}
\usage{
calc_variables_stem_level(data_stems, height_model)
}
\arguments{
\item{data_stems}{dataframe on stems (shoots and trees) as given from the
function compose_stem_data()}

\item{height_model}{dataframe with \code{model} containing 'exp' or 'ln',
coefficients \code{P1} and \code{P2} to calculate height model for each combination of
\code{species}, \code{forest_reserve}, \code{period} and \code{plottype}. Height models in .xlsx
generated by \code{Fieldmap} can be grouped in a dataframe using function
\code{load_height_models()}}
}
\value{
Dataframe with ...
}
\description{
This function calculates additional variables based on measurements, such as
\itemize{
\item \code{calc_height_m}: calculated height based on \code{dbh_mm} and a species
specific diameter-height model
\item \code{basal_area_m2}
\item \code{vol_bole_m3}: calculated based on \code{dbh_mm}, \code{calc_height_m} and
species specific tariffs
\item \code{vol_crown_m3}: calculated based on \code{dbh_mm} and
species specific tariffs
\item \code{vol_tot_m3}: sum of \code{vol_bole_m3} and \code{vol_crown_m3}
\item \code{basal_area_alive_m2_ha}
\item \code{basal_area_dead_m2_ha}
\item \code{vol_alive_m3_ha}
\item \code{vol_dead_standing_m3_ha}
\item \code{vol_bole_alive_m3_ha}
\item \code{vol_bole_dead_m3_ha}
}
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")

data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
# omit argument 'example_dataset = TRUE' below to use all height models
height_model <- load_height_models(example_dataset = TRUE)
calc_variables_stem_level(data_stems, height_model)

}
