% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_variables_tree_level.R
\name{calc_variables_tree_level}
\alias{calc_variables_tree_level}
\title{Calculate additional variables on tree level}
\usage{
calc_variables_tree_level(data_dendro, data_shoots, data_stems, height_model)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species, alive_dead, decaystage, vol_tot_m3, basal_area_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}

\item{data_shoots}{dataframe on shoots as given from the function load_data_shoots()}

\item{data_stems}{dataframe on stems (shoots and trees) as given from the function compose_stem_data()}

\item{height_model}{dataframe with `model` containing 'exp' or 'ln',
coeficients `P1` and `P2` to calculate height model for each combination of
`species`, `forest_reserve`, `period` and `plot_type`. Height models in .xlsx
generated by Fieldmap can be grouped in a dataframe using function
`load_height_models()`}
}
\value{
Dataframe with ...
}
\description{
This function calculates additional variables based on measurements, such as
\itemize{
 \item tree_number: the number of shoots in the tree (> 1 when coppice)
 \item individual: true if secondary shoot is present, false for individual tree and coppice shoot
 \item calc_height_m: calculated height based on `dbh_mm` and a model
 \item basal_area_m2
 \item vol_tot_m3
 \item vol_stem_m3
 \item vol_crown_m3
 \item dbh_mm (based on average for coppice trees)
 \item decaystage (based on average for coppice trees)
}
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_shoots <-
  load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models("C:/bosreservaten/Hoogtemodellen/")
calc_variables_tree_level(data_dendro, data_shoots, data_stems, height_model)
}

}
