% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deadw_decay_plot_species.R
\name{calculate_deadw_decay_plot_species}
\alias{calculate_deadw_decay_plot_species}
\title{aggregate parameters by decay stage, plot, tree species and year}
\usage{
calculate_deadw_decay_plot_species(data_deadwood)
}
\arguments{
\item{data_deadwood}{dataframe on logs with variables plot_id, plottype,
date_dendro, species, decaystage, calc_volume_m3, period and year
(output of function load_data_deadwood())}
}
\value{
dataframe with columns plot, year, tree_species, decaystage,
vol_log_m3_ha
}
\description{
This function calculates for each plot, tree species and year the volume
logs per hectare and per decay stage.
}
\examples{
\dontrun{
#change path before running
data_deadwood <-
  load_data_deadwood("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_deadw_decay_plot_species(data_deadwood)
}

}
