% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendro_plot_year.R
\name{calculate_dendro_plot_year}
\alias{calculate_dendro_plot_year}
\title{aggregate parameters by plot and year}
\usage{
calculate_dendro_plot_year(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables ...}
}
\value{
dataframe with columns plot, year, number_of_tree_species_ha, number_of_trees_ha, basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot and year some values per hectare: number of tree species, number of trees, basal area, and volume.
}
