% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendrometry.R
\name{calculate_dendrometry}
\alias{calculate_dendrometry}
\title{aggregates parameters on dendrometry of trees}
\usage{
calculate_dendrometry(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables ...}
}
\value{
List of dataframes that are mentioned in the above description
}
\description{
This function makes aggregations of individual tree measures on the levels of
\itemize{
 \item plot and year
 \item plot, tree species and year
 \item plot (comparison between different years)
 \item plot and tree species (comparison between different years)
 \item individual trees: statuses in different years
}
}
