% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diam_plot.R
\name{calculate_diam_plot}
\alias{calculate_diam_plot}
\title{aggregate parameters by diameter class, plot and year}
\usage{
calculate_diam_plot(data_stems_calc, data_deadwood, plotinfo)
}
\arguments{
\item{data_stems_calc}{dataframe on stem level measurements with variables
plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species,
alive_dead, decaystage, period, year, subcircle, plotarea_ha,...
(output of function calc_variables_stem_level())}

\item{data_deadwood}{dataframe on logs with variables plot_id, plottype,
date_dendro, species, decaystage, calc_volume_m3, period and year
(output of function load_data_deadwood())}

\item{plotinfo}{dataframe on surveyed plots with variables plot_id, plottype,
forest_reserve, survey_trees, survey_deadw, period and year_dendro
(output of function load_plotinfo())}
}
\value{
dataframe with columns plot, year, dbh_class_5cm, number_of_trees_ha,
basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot, year and diameter class some values
per hectare: number of stems, basal area and volume of standing trees
(for coppice based on data on shoot level),
and volume of logs (= lying deadwood).
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")

data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models()
data_stems_calc <- calc_variables_stem_level(data_stems, height_model)
data_deadwood <- load_data_deadwood(path_to_fieldmapdb)
plotinfo <- load_plotinfo(path_to_fieldmapdb)
calculate_diam_plot(data_stems_calc, data_deadwood, plotinfo)

}
