% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diam_plot_species_year.R
\name{calculate_diam_plot_species_year}
\alias{calculate_diam_plot_species_year}
\title{aggregate parameters by diameter class, plot, tree species and year}
\usage{
calculate_diam_plot_species_year(data_dendro, data_stems)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, tree_measure_id, date_dendro, DBH_mm, Height_m, species, AliveDead, decaystage, Adjust_Vol_tot_m3, AdjustBasalArea_m2, period, OldID, year, plottype, plotarea_ha,... (output of function load_data_dendrometry())}

\item{data_stems}{data on stems (shoots and trees) as given from the function compose_stem_data()}
}
\value{
dataframe with columns plot, year, tree_species, DBHClass_5cm, basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot, tree species, year and diameter class some values per hectare: number of trees, basal area, and volume.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_shoots <-
  load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_stems <- compose_stem_data(data_dendro, data_shoots)
calculate_diam_plot_species_year(data_dendro, data_stems)
}

}
