% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diam_plot_year.R
\name{calculate_diam_plot_year}
\alias{calculate_diam_plot_year}
\title{aggregate parameters by diameter class, plot and year}
\usage{
calculate_diam_plot_year(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, tree_measure_id, date_dendro, DBH_mm, Height_m, species, AliveDead, decaystage, Adjust_Vol_tot_m3, AdjustBasalArea_m2, period, OldID, year, plottype, plotarea_ha,... (output of function load_data_dendrometry())}
}
\value{
dataframe with columns plot, year, DBHClass_5cm, number_of_trees_ha, basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot, year and diamter class some values per hectare: number of trees, basal area, and volume.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_diam_plot_year(data_dendro)
}

}
