% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_core_area_species.R
\name{calculate_regeneration_core_area_species}
\alias{calculate_regeneration_core_area_species}
\title{aggregate regeneration parameters by plot, species and year}
\usage{
calculate_regeneration_core_area_species(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
plot_id, plottype, subplot_id, height_class, species, nr_of_regeneration,
rubbing_damage_number, period, year, subcircle, plotarea_ha,
min_number_of_regeneration and max_number_of_regeneration.}
}
\value{
dataframe with columns plot, species, year,
nr_of_subplots_with_regeneration, perc_subplots_with_regeneration,
mean_number_established_ha,
lci_number_established_ha, uci_number_established_ha,
mean_number_seedlings_ha,
lci_number_seedlings_ha, uci_number_seedlings_ha,
mean_rubbing_damage_perc_established, lci_rubbing_damage_perc_established,
uci_rubbing_damage_perc_established,
mean_rubbing_damage_perc_seedlings, lci_rubbing_damage_perc_seedlings,
uci_rubbing_damage_perc_seedlings,
approx_nr_established_ha, approx_nr_seedlings_ha,
approx_rubbing_damage_perc_established, approx_rubbing_damage_perc_seedlings.
}
\description{
This function calculates for each plot, species and year the number of
seedlings and established regeneration per ha (or interval with mean
and confidence interval using a log transformation), the number and
percentage of subplots in which the species is regenerating and
the approximate rubbing damage percentage per hectare.
This calculation is designed for core areas,that consist of different
subplots.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration_CA <-
  load_data_regeneration(path_to_fieldmapdb, plottype = "CA")
calculate_regeneration_core_area_species(data_regeneration_CA)

}
