% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot.R
\name{calculate_regeneration_plot}
\alias{calculate_regeneration_plot}
\title{calculate species number by plot and year}
\usage{
calculate_regeneration_plot(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables ...}
}
\value{
dataframe with columns plot, year and number_of_tree_species
}
\description{
This function calculates for each plot and year the number of species, total
number of seedlings and established regeneration (or interval with mean
and confidence interval) and rubbing damage percentage for regeneration
(for all height classes together).
For core area plots, these variables are calculated for each subplot.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_regeneration <-
  load_data_regeneration("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_regeneration_plot(data_regeneration)
}

}
