% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot_height.R
\name{calculate_regeneration_plot_height}
\alias{calculate_regeneration_plot_height}
\title{calculate species number by plot, tree height class and year}
\usage{
calculate_regeneration_plot_height(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables ...}
}
\value{
dataframe with columns plot_id, year, period, height_class,
number_of_tree_species, rubbing_damage_perc,
nr_of_regeneration (in case exact numbers are observed),
mean_number_of_regeneration (in case intervals are observed),
lci_number_of_regeneration and uci_number_of_regeneration.
}
\description{
This function calculates for each plot, tree height class and year
the number of species, total number of regeneration (or interval with mean
and confidence interval) and rubbing damage percentage for regeneration.
For core area plots, these variables are calculated for each subplot.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_regeneration <-
  load_data_regeneration("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_regeneration_plot_height(data_regeneration)
}

}
