% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_regeneration_plot_height_species_diff.R
\name{calculate_regeneration_plot_height_species_diff}
\alias{calculate_regeneration_plot_height_species_diff}
\title{aggregate parameters by plot, height class and species}
\usage{
calculate_regeneration_plot_height_species_diff(
  regeneration_by_plot_height_species
)
}
\arguments{
\item{regeneration_by_plot_height_species}{dataframe with values for each plot, height class, species and year, which is the result of the calculation by function calculate_regeneration_plot_height_species() and can be retrieved from forresdat}
}
\value{
dataframe with columns plot, height class, species and number_of_trees_ha_diff
}
\description{
This function compares for each plot, height class and species the differences between years for the number of individuals per hectare. It gives results for differences between subsequent measures (based on 'period') and between the last and the first measure.
}
\examples{
\dontrun{
#change path before running
regeneration_by_plot_height_species <-
  read_git(tablename = "regeneration_by_plot_height_species", repo_path = "C:/gitrepo/forresdat")
calculate_regeneration_plot_height_species_diff(regeneration_by_plot_height_species)
}

}
