% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_regeneration_plot_height_species_year.R
\name{calculate_regeneration_plot_height_species_year}
\alias{calculate_regeneration_plot_height_species_year}
\title{aggregate parameters by plot, tree height class, species and year}
\usage{
calculate_regeneration_plot_height_species_year(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables ...}
}
\value{
dataframe with columns plot, year, height_class, species and number_of_trees_ha
}
\description{
This function calculates for each plot, tree height class, species and year the number of trees per hectare for regeneration.
}
\examples{
\dontrun{
#change path before running
data_regeneration <-
  load_data_regeneration("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_regeneration_plot_height_species_year(data_regeneration)
}

}
