% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot_year.R
\name{calculate_regeneration_plot_year}
\alias{calculate_regeneration_plot_year}
\title{calculate species number by plot and year}
\usage{
calculate_regeneration_plot_year(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables ...}
}
\value{
dataframe with columns plot, year and number_of_tree_species
}
\description{
This function calculates for each plot and year the number of species for generation (for all height classes together).
}
