% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vegetation_plot_species.R
\name{calculate_vegetation_plot_species}
\alias{calculate_vegetation_plot_species}
\title{give percentage presence of species in subplots}
\usage{
calculate_vegetation_plot_species(data_vegetation)
}
\arguments{
\item{data_vegetation}{dataframe on vegetation with variables ...}
}
\value{
dataframe with columns plot, species, year and perc_of_subplots
}
\description{
This function calculates for each plot, species and year the percentage of subplots in which the species is present
}
\examples{
\dontrun{
#change path before running
library(dplyr)
data_vegetation_CA <-
  load_data_vegetation("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb") \%>\%
  filter(Plottype == 30)
calculate_vegetation_plot_species(data_vegetation_CA)
}

}
