% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vegetation_subplot.R
\name{calculate_vegetation_subplot}
\alias{calculate_vegetation_subplot}
\title{aggregate vegetation parameters by subplot and year}
\usage{
calculate_vegetation_subplot(data_vegetation)
}
\arguments{
\item{data_vegetation}{dataframe on vegetation with variables ...}
}
\value{
dataframe with columns plot, subplot, year and number_of_species
}
\description{
This function calculates for each subplot and year the the number of species and cumulated herb layer coverage in the vegetation layer.  This calculation is designed for core areas, that consist of different subplots.  In other plot types, this function gives the same result as calculate_vegetation_plot().
}
\examples{
\dontrun{
#change path before running
data_vegetation <-
  load_data_vegetation(
    "C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb",
    plottype = "Core area"
  )
calculate_vegetation_subplot(data_vegetation)
}

}
