% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_fmdb.R
\name{check_data_fmdb}
\alias{check_data_fmdb}
\title{check tables from fieldmap database for inconsistencies}
\usage{
check_data_fmdb(database, forest_reserve = "all")
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{forest_reserve}{name of forest reserve for which the records in the
database should be checked (defaults to "all")}
}
\value{
Dataframe with inconsistent data with layer, ID's and
additional columns
`aberrant_field` (which column is wrong) and `anomaly` (what is wrong with
the input)
}
\description{
This function retrieves the important fields of tables `Trees`, `Shoots`,
`Deadwood`, `Regeneration`, `Regspecies`, `Vegetation`, `Herblayer`, `Plots`
and `Plotdetails` (of all periods) from the given database and
checks for missing data or wrong input.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
check_data_fmdb(path_to_fieldmapdb)
check_data_fmdb(path_to_fieldmapdb, forest_reserve = "Everzwijnbad")

}
