% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_shoots.R
\name{check_data_shoots}
\alias{check_data_shoots}
\title{check table \code{Shoots} from \code{Fieldmap} database for inconsistencies}
\usage{
check_data_shoots(database, forest_reserve = "all")
}
\arguments{
\item{database}{name of \code{Fieldmap}/Access database (with specific \code{Fieldmap}
structure) including path}

\item{forest_reserve}{name of forest reserve for which the records in the
database should be checked (defaults to "all")}
}
\value{
Dataframe with inconsistent data with ID's and additional columns
\code{aberrant_field} (which column is wrong) and \code{anomaly} (what is wrong with
the input)
}
\description{
This function retrieves the important fields of table \code{Shoots} (of all
periods) from the given database and checks for missing data or wrong input.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
check_data_shoots(path_to_fieldmapdb)
check_data_shoots(path_to_fieldmapdb, forest_reserve = "Everzwijnbad")

}
