% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_trees.R
\name{check_data_trees}
\alias{check_data_trees}
\title{check table trees from fieldmap database for inconsistencies}
\usage{
check_data_trees(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap structure) including path}
}
\value{
Dataframe with inconsistent data (FOR NOW ALL DATA WITH EXTRA COLUMN 'PROBLEM')
}
\description{
STILL UNDER CONSTRUCTION!!!
This function retrieves the important fields of table Trees (and Trees_2eSET) from the given database and checks for lacking data or wrong input.
}
\examples{
\dontrun{
#change path before running
check_data_trees("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
