% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_vegetation.R
\name{check_data_vegetation}
\alias{check_data_vegetation}
\title{check table `Vegetation` from fieldmap database for inconsistencies}
\usage{
check_data_vegetation(database, forest_reserve = "all")
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{forest_reserve}{name of forest reserve for which the records in the
database should be checked (defaults to "all")}
}
\value{
Dataframe with inconsistent data with ID's and additional columns
`aberrant_field` (which column is wrong) and `anomaly` (what is wrong with
the input)
}
\description{
This function retrieves the important fields of table `Vegetation`
(of all periods) from the given database and
checks for missing data or wrong input.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
check_data_vegetation(path_to_fieldmapdb)
check_data_vegetation(path_to_fieldmapdb, forest_reserve = "Everzwijnbad")

}
