% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_periods.R
\name{compare_periods}
\alias{compare_periods}
\title{compare parameters between periods (years that parameters are measured)}
\usage{
compare_periods(dataset, measure_vars)
}
\arguments{
\item{dataset}{dataframe with values for each period, plot and year, in addition to grouping variables and measure_vars}

\item{measure_vars}{column names of variables that should be compared between periods (including year)}
}
\value{
dataframe with columns plot, year_diff, n_years, grouping variables and differences between periods for each column of measure_vars
}
\description{
This function compares for each plot (and other provided variables) the differences between periods/years for the column names given in parameter measure_vars. It gives results for differences between subsequent measures (based on 'period') and between the last and the first measure. All column names of the dataset that are not added to parameter measure_vars, are considered as grouping variables, except for period. If the result is not as expected, please verify that the dataset only consists of grouping variables, variables added to measure_vars and period.
}
\examples{
\dontrun{
#change path before running
treenr_by_plot <-
  read_git(tablename = "dendro_by_plot", repo_path = "C:/gitrepo/forresdat") \%>\%
  select(period, year, plot_id, number_of_tree_species, number_of_trees_ha) \%>\%
  distinct()
calculate_dendro_plot_diff(treenr_by_plot, c("year", "number_of_tree_species", "number_of_trees_ha"))
}

}
