% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.R
\name{connect_to_database}
\alias{connect_to_database}
\title{connect to fieldmap database}
\usage{
connect_to_database(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}
}
\value{
DBI connection that can be used to connect to the database
}
\description{
This helper function returns a connection to the given database (path).
Reason for this function is to avoid repetition of this information.
}
