% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diam_statistics_forres_species.R
\name{create_diam_statistics_forres_species}
\alias{create_diam_statistics_forres_species}
\title{Give diameter statistics by forest reserve, species and year}
\usage{
create_diam_statistics_forres_species(data_stems)
}
\arguments{
\item{data_stems}{dataframe on stems (shoots and trees) as given from the function compose_stem_data()}
}
\value{
dataframe with columns forest_reserve, species, year and measures on diameter distribution (min, max, mean, median, Q1, Q3)
}
\description{
This function calculates the diameter distribution on the level of forest reserve, species and year
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_shoots <-
  load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_stems <- compose_stem_data(data_dendro, data_shoots)
create_diam_statistics_forres_species(data_stems)
}

}
