% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statistics_dendro_diam_forres.R
\name{create_statistics_dendro_diam_forres}
\alias{create_statistics_dendro_diam_forres}
\title{Calculate statistics by forest reserve, diameter and year}
\usage{
create_statistics_dendro_diam_forres(dendro_by_diam_plot)
}
\arguments{
\item{dendro_by_diam_plot}{data on the level of plot, diameter and year, e.g. table dendro_by_diam_plot in git repository forresdat or the result of function calculate_diam_plot()}
}
\value{
dataframe with columns forres, year, number_of_trees_ha, basal_area_m2_ha, volume_m3_ha
}
\description{
This function shows data from the git-repository forresdat on the level of plot and year some values per diameter class and hectare: number of trees, basal area, and volume.
}
\examples{
\dontrun{
#change path before running
dendro_by_diam_plot <-
  read_git(tablename = "dendro_by_diam_plot", repo_path = "C:/gitrepo/forresdat")
create_statistics_dendro_diam_forres(dendro_by_diam_plot)
}

}
