% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unique_tree_id.R
\name{create_unique_tree_id}
\alias{create_unique_tree_id}
\title{create_unique_tree_id for each individual tree over different years}
\usage{
create_unique_tree_id(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id,
plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species,
alive_dead, decaystage, period, OldID, year, subcircle, plotarea_ha,...
(output of function load_data_dendrometry())}
}
\value{
a dataset with 1 record per tree measurement, containing the given
data of each tree in different years (= data_dendro) and a link
to a unique tree_id.
}
\description{
This function creates a unique ID for each tree, that allows to group
(f.e. by use of `make_table_wide()`) all given information on the life stages
of an individual tree during different measures.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
create_unique_tree_id(data_dendro)
}

}
