% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_access_to_forresdat.R
\name{from_access_to_forresdat}
\alias{from_access_to_forresdat}
\title{copy table(s) from access database to git repository forresdat}
\usage{
from_access_to_forresdat(
  database,
  tables,
  repo_path,
  metadata_path,
  push = FALSE,
  strict = TRUE,
  branch = "develop"
)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{tables}{vector with table names of tables that should be moved}

\item{repo_path}{name and path of local forresdat repository in which
results/tables should be saved}

\item{metadata_path}{path including .xlsx file in which the metadata are
stored}

\item{push}{push commits directly to the remote on Github?
Default is FALSE (no). (This option can only be used with SSH.)}

\item{strict}{keep default TRUE to update data without structural changes,
change to FALSE only if tables are structurally changed
(e.g. additional column, change in sorting order,...)}

\item{branch}{branch from repository forresdat to which the new version
should be committed.
Default is 'develop'.}
}
\value{
No value is returned, the tables are saved in the git repository.
}
\description{
This function loads one or more tables from the access database
(or an SQLite database) and saves them in the git repository forresdat.
Table names in camelcase in the database are renamed to snakecase before
saving in forresdat.
}
\examples{
\dontrun{
#make a local clone of forresdat and change path before running
library(forrescalc)
# add path to your local clone of forresdat
path_to_forresdat <- "xxx/forresdat"
# if you don't have a local clone yet, make it:
git2r::clone("https://github.com/inbo/forresdat.git", path_to_forresdat)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
# add path to metadata here
temp <- tempfile(fileext = ".xlsx")
dl <- googledrive::drive_download(
         googledrive::as_id("12x2H9lp86R-AFPdN2JXB9nqwJ2_A6PF6"),
         path = temp, overwrite = TRUE
       )

from_access_to_forresdat(
  database = path_to_fieldmapdb,
  tables = c("qCoverHerbs", "qtotalCover"),
  repo_path = path_to_forresdat,
  metadata_path = temp
)
}

}
