% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_access_to_git.R
\name{from_access_to_git}
\alias{from_access_to_git}
\title{copy table(s) from access db to git repository}
\usage{
from_access_to_git(database, tables, repo_path)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap structure) including path}

\item{tables}{vector with table names of tables that should be moved}

\item{repo_path}{name and path of local git repository in which results should be saved}
}
\value{
No value is returned, the tables are saved in the git repository.
}
\description{
This function loads one or more tables from the access database and saves them in a git repository.
}
