% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_diamclass_5cm.R
\name{give_diamclass_5cm}
\alias{give_diamclass_5cm}
\title{Calculate the diameter class of diameter measure data}
\usage{
give_diamclass_5cm(diameterdata)
}
\arguments{
\item{diameterdata}{vector with diameter data in millimetre}
}
\value{
vector with factors in diameter classes of 5 cm
}
\description{
This function returns a factor with diameter classes of 5 cm for a given
vector with diameter data in mm.
}
\examples{
library(forrescalc)
give_diamclass_5cm(c(80, 1512, 2222))

}
