% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_deadwood.R
\name{load_data_deadwood}
\alias{load_data_deadwood}
\title{retrieves data on logs from fieldmap database}
\usage{
load_data_deadwood(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap structure) including path}
}
\value{
Dataframe with data on logs
}
\description{
This function queries the given database to retrieve data on deadwood (logs) (ready for use in calculate_dendrometry function).
}
