% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_herblayer.R
\name{load_data_herblayer}
\alias{load_data_herblayer}
\title{retrieve species specific vegetation data from fieldmap database}
\usage{
load_data_herblayer(database, plottype = NA, forest_reserve = NA)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{plottype}{possibility to select only data for a certain plot type, e.g. 'circular plot'
or 'core area' (the default NA means that data from all plots are retrieved)}

\item{forest_reserve}{possibility to select only data for 1 forest reserve
by giving the name of the forest reserve (the default NA means that data
from all plots are retrieved)}
}
\value{
Dataframe with vegetation data on the species level ('herb layer')
}
\description{
This function queries the given database to retrieve data on vegetation (ready for use in calculate_vegetation function).
}
\examples{
\dontrun{
#change path before running
load_data_herblayer("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
