% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_regeneration.R
\name{load_data_regeneration}
\alias{load_data_regeneration}
\title{retrieve regeneration data from fieldmap database}
\usage{
load_data_regeneration(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap structure) including path}
}
\value{
Dataframe with regeneration data
}
\description{
This function queries the given database to retrieve data on regeneration (ready for use in calculate_regeneration function).
}
\examples{
\dontrun{
#change path before running
load_data_regeneration("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
