% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_height_models.R
\name{load_height_models}
\alias{load_height_models}
\title{retrieve height model data from xlsx files}
\usage{
load_height_models(path_to_height_models)
}
\arguments{
\item{path_to_height_models}{path to folder where height models are stored}
}
\value{
Dataframe with height model data
}
\description{
This function groups the information on height models from the `.xlsx` files
in the given folder together in one dataframe.
}
\examples{
#change path before running
library(forrescalc)
# (add path to your height models here)
path_to_height_models <-
  system.file("example/height_models", package = "forrescalc")
load_height_models(path_to_height_models)

}
