% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_plotdata.R
\name{load_plotdata}
\alias{load_plotdata}
\title{retrieve generic plot data from fieldmap database}
\usage{
load_plotdata(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}
}
\value{
Dataframe with columns plot_id, plottype and forest_reserve
}
\description{
This function queries the given database to retrieve additional data on plots
to save in forresdat and link with the datasets that are saved there.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
load_plotdata("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
