% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_wide.R
\name{make_table_wide}
\alias{make_table_wide}
\title{change dataframe from long to wide format}
\usage{
make_table_wide(table_long, column_to_repeat, columns_for_comparison)
}
\arguments{
\item{table_long}{dataframe with data in long format. It is important that
all columns that are not mentioned in the variables \code{column_to_repeat}
or \code{columns_for_comparison}, are grouping variables that have the same
value for each \code{column_to_repeat} (see 2nd example).}

\item{column_to_repeat}{name of the column of which the values have to be
added to the column headings}

\item{columns_for_comparison}{(vector with) name(s) of the column(s) you
want to repeat for each value of \code{column_to_repeat}}
}
\value{
the dataframe in long format
}
\description{
This function changes a dataframe from long to wide format, e.g. to show data
from 2 different periods in different columns.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
library(dplyr)
table_long <-
  read_forresdat(
    tablename = "dendro_by_plot_species",
    repo_path = "C:/Users/els_lommelen/Documents/gitrepos/forresdat"
  ) \%>\%
  filter(plot_id < 110) \%>\%
  select(plot_id, species, period, number_of_trees_ha, vol_alive_m3_ha)
table_wide <-
  make_table_wide(
    table_long, column_to_repeat = "period",
    columns_for_comparison = c("number_of_trees_ha", "vol_alive_m3_ha"))
#if number_of_trees_ha is not mentioned in columns_for_comparison, it is
#considered as a grouping variable while it has different values for each
#period.
#This gives an unwanted result with still many rows and a lot of NA values:
table_wide <-
  make_table_wide(table_long, column_to_repeat = "period",
                  columns_for_comparison = c("vol_alive_m3_ha"))
}

}
