% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_database.R
\name{query_database}
\alias{query_database}
\title{query all periods from fieldmap database}
\usage{
query_database(
  database,
  query,
  selection = "",
  add_fields = "",
  conjunction = "",
  n_periods = 3L
)
}
\arguments{
\item{database}{name of \code{Fieldmap}/Access database (with specific \code{Fieldmap}
structure) including path}

\item{query}{query that is adapted to be used in different periods,
in which the following parameters can be used: \itemize{
\item \verb{\\\%1$d} : replaced by the period number,
\item \verb{\\\%2$s} : replaced by an empty string in period 1 and by \verb{_?eSet}
in all other periods (with ? the period)
\item \verb{\\\%3$s} : replaced by the string given in argument \code{selection}
\item \verb{\\\%4$s} : replaced by the string given in argument \code{add_fields}
\item \verb{\\\%5$s} : replaced by the string given in argument \code{conjunction}
}}

\item{selection}{string that will be added to \verb{\\\%3$s} in the query}

\item{add_fields}{string that will be added to \verb{\\\%4$s} in the query}

\item{conjunction}{string that will be added to \verb{\\\%5$s} in the query}

\item{n_periods}{highest period number that is present in the database
(change default here when a new period is added to the database)}
}
\value{
Dataframe containing the result of the query and an additional column
period
}
\description{
This helper function retrieves data of all periods from the given database
by using the given query and adding in the period dependent parameters.
Reason for this function is to avoid repetition in the load functions.
}
