% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_git.R
\name{read_git}
\alias{read_git}
\title{load tables from git repository}
\usage{
read_git(tablename, repo_path)
}
\arguments{
\item{tablename}{name of the table that should be read}

\item{repo_path}{name and path of local git repository in which results should be saved}
}
\value{
A dataframe with the specified table
}
\description{
This function reads a table in git2rdata format from a git repository.
}
\examples{
\dontrun{
#change path before running
read_git(tablename = "dendro_status_tree", repo_path = "C:/gitrepo/forresdat")
}

}
