% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_last_commit_git.R
\name{remove_last_commit_git}
\alias{remove_last_commit_git}
\title{remove last local change from git repository}
\usage{
remove_last_commit_git(repo_path)
}
\arguments{
\item{repo_path}{name and path of local git repository in which last
commits should be removed}
}
\value{
A dataframe with the specified table
}
\description{
This function removes the last commit from the active branch of the
specified git repository. ONLY USE THIS FUNCTION IF YOUR COMMIT IS
NOT YET PUSHED TO THE REMOTE!!! This function is meant for users that
are not familiar with Git to easily remove an automatically generated
commit in forresdat after they discoved mistakes in it.
}
\examples{
\dontrun{
#change paths before running
library(forrescalc)
# add path to your local clone of forresdat
path_to_forresdat <- "xxx/forresdat"

# only run this after writing a commit with `save_results_git()` or
# `from_access_to_git()` that has not yet been pushed to Github!
remove_last_commit_git(repo_path = path_to_forresdat)
}

}
