% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_table_forresdat.R
\name{remove_table_forresdat}
\alias{remove_table_forresdat}
\title{remove table(s) from data package forresdat}
\usage{
remove_table_forresdat(tables, repo_path, push = FALSE, branch = "develop")
}
\arguments{
\item{tables}{vector with table names of tables that should be removed}

\item{repo_path}{name and path of local forresdat repository in which
results/tables should be saved}

\item{push}{push commits directly to the remote on github?
Default is FALSE (no). (This option can only be used with SSH.)}

\item{branch}{branch from repository forresdat to which the new version
should be committed.
Default is 'develop'.}
}
\value{
No value is returned, the tables are removed from the git repository.
}
\description{
This function removes one or more tables from the data package forresdat
by making a commit on a local clone of the git repository.
While removing the table(s), it also updates the metadata (`.json` file)
}
\examples{
\dontrun{
#make a local clone of forresdat and change path before running
library(forrescalc)
# add path to your local clone of forresdat
path_to_forresdat <- "xxx/forresdat"
# if you don't have a local clone yet, make it:
git2r::clone("https://github.com/inbo/forresdat.git", path_to_forresdat)

remove_table_forresdat(
  tables = c("qCoverHerbs", "qtotalCover"),
  repo_path = path_to_forresdat
)
}

}
