% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results_access.R
\name{save_results_access}
\alias{save_results_access}
\title{save results of calculations in Access database}
\usage{
save_results_access(results, database)
}
\arguments{
\item{results}{results from calculations in package forrescalc as a named list}

\item{database}{name of (empty) Access database including path in which results should be saved}
}
\value{
No value is returned, data are saved in the specified database
}
\description{
This function saves the results from calculations in the forrescalc package (or any other named list with dataframes) in an Access database.  List item names will be used to name each of the tables, which contain as a content the different dataframes.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
result_dendro <- calculate_dendrometry(data_dendro)
save_results_access(result = result_dendro, database = "C:/db/testdb.accdb")
}

}
