% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results_csv.R
\name{save_results_csv}
\alias{save_results_csv}
\title{save results of calculations as \code{.csv}-files}
\usage{
save_results_csv(results, output_dir)
}
\arguments{
\item{results}{results from calculations in package forrescalc as a named
list}

\item{output_dir}{name of output folder including path in which results
should be saved}
}
\value{
No value is returned, data are saved in the specified folder
}
\description{
This function saves the results from calculations in the forrescalc package
(or any other named list with dataframes) in a predefined folder.
List item names will be used to name each of the tables, which contain as
a content the different dataframes.
}
\examples{
library(forrescalc)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
path_to_plotlevel_csv <- getwd()
data_regeneration <- load_data_regeneration(database = path_to_fieldmapdb)
regeneration <- calculate_regeneration(data_regeneration)
save_results_csv(results = regeneration, output_dir = path_to_plotlevel_csv)

files <- list.files()
files <- files[grepl("^regeneration_by.*\\\\.csv$", files)]
file.remove(files)

}
