% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_for_DAS_indicator.R
\name{select_for_DAS_indicator}
\alias{select_for_DAS_indicator}
\title{select relevant plots/reserves to calculate DAS indicator}
\usage{
select_for_DAS_indicator(data_to_select, grouping_vars)
}
\arguments{
\item{data_to_select}{dataframe with at least all grouping_vars and the variables dbh_mm, species, basal_area_alive_m2_ha, and as records preferably only living trees (they are not filtered out here but they should not be in here to meet the requirements of the DAS indicator)}

\item{grouping_vars}{vector with variables that should be grouped on during the selection steps.}
}
\value{
dataframe with the grouping vars in which records are removed that do not meet the above described criteria
}
\description{
This function does the first step(s) of the calculation of the DAS indicator, namely selecting the relevant forest reserves or plots based on the following criteria:

\itemize{
 \item test for development stages: average diameter (DBH) between 200 and 700 mm
 \item test for desired tree species: > 5 \% of the proportion of the basal area should contain trees from the species groups beech, oak, ash/maple or pioneers (birch, willow, esp or rowan)
 \item exclude homogeneous stands: basal area proportion of each tree species group < 98 \%
}
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
select_for_DAS_indicator(data_dendro)
}

}
