% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_intervals.R
\name{sum_intervals}
\alias{sum_intervals}
\title{Sum up intervals including a confidence interval}
\usage{
sum_intervals(var_min, var_max, transformation = NA, na_rm = FALSE)
}
\arguments{
\item{var_min}{minimum of intervals to be summed up}

\item{var_max}{maximum of intervals to be summed up}

\item{transformation}{"log" if log transformation has to be performed on data}

\item{na_rm}{should NA in var_min and var_max be ignored?}
}
\value{
dataframe with one record with columns n_obs, sum, lci (lower limit
of confidence interval) and uci (upper limit of confidence interval)
}
\description{
This internal helper function is used to sum up interval data, hereby
respecting the confidence interval
}
\examples{
#see calculate_regeneration_plot_height()

}
