% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_status.R
\name{summarise_status}
\alias{summarise_status}
\title{give the statuses of each individual in different years}
\usage{
summarise_status(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables ...}
}
\value{
a dataset with 1 record for each tree, containing the decay stage in different years
}
\description{
This function groups all information on the life stages of an individual tree during different measures.
}
