% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_variables_stem_level.R
\name{calc_variables_stem_level}
\alias{calc_variables_stem_level}
\title{Calculate additional variables on stem level}
\usage{
calc_variables_stem_level(data_stems, height_model)
}
\arguments{
\item{data_stems}{dataframe on stems (shoots and trees) as given from the
function compose_stem_data()}

\item{height_model}{dataframe with `model` containing 'exp' or 'ln',
coeficients `P1` and `P2` to calculate height model for each combination of
`species`, `forest_reserve`, `period` and `plottype`. Height models in .xlsx
generated by Fieldmap can be grouped in a dataframe using function
`load_height_models()`}
}
\value{
Dataframe with ...
}
\description{
This function calculates additional variables based on measurements, such as
\itemize{
 \item calc_height_m: calculated height based on `dbh_mm` and a species
 specific diameter-height model
 \item basal_area_m2
 \item vol_bole_m3: calculated based on `dbh_mm`, `calc_height_m` and
 species specific tariffs
 \item vol_crown_m3: calculated based on `dbh_mm` and
 species specific tariffs
 \item vol_tot_m3: sum of `vol_bole_m3` and `vol_crowwn_m3`
 \item basal_area_alive_m2_ha
 \item basal_area_dead_m2_ha
 \item vol_alive_m3_ha
 \item vol_dead_standing_m3_ha
 \item vol_bole_alive_m3_ha
 \item vol_bole_dead_m3_ha
}
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("database/mdb_bosres.sqlite", package = "forrescalc")
data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models("C:/bosreservaten/Hoogtemodellen/")
calc_variables_stem_level(data_stems, height_model)

}
