% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_variables_tree_level.R
\name{calc_variables_tree_level}
\alias{calc_variables_tree_level}
\title{Calculate additional variables on tree level}
\usage{
calc_variables_tree_level(data_dendro, data_stems_calc)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id,
plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species,
alive_dead, decaystage, period, OldID, year, subcircle, plotarea_ha,...
(output of function load_data_dendrometry())}

\item{data_stems_calc}{dataframe on stem level measurements with variables
plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species,
alive_dead, decaystage, period, year, subcircle, plotarea_ha,...
(output of function calc_variables_stem_level())}
}
\value{
Dataframe with ...
}
\description{
This function calculates additional variables based on measurements, such as
\itemize{
 \item nr_of_stems: the number of shoots in the tree
 (= 1 for an individual tree; >= 1 when coppice)
 \item individual: true for individual tree or coppice,
 false if record is a secondary shoot
 \item calc_height_m: calculated height based on `dbh_mm` and
 a species specific diameter-height model
 \item basal_area_m2
 \item vol_bole_m3: calculated based on `dbh_mm`, `calc_height_m` and
 species specific tariffs
 \item vol_crown_m3: calculated based on `dbh_mm` and
 species specific tariffs
 \item vol_tot_m3: sum of `vol_bole_m3` and `vol_crowwn_m3`
 \item dbh_mm (based on average for coppice trees)
 \item decaystage (based on average for coppice trees)
 \item basal_area_alive_m2_ha
 \item basal_area_dead_m2_ha
 \item vol_alive_m3_ha
 \item vol_dead_standing_m3_ha
 \item vol_bole_alive_m3_ha
 \item vol_bole_dead_m3_ha
}
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models("C:/bosreservaten/Hoogtemodellen/")
data_stems_calc <- calc_variables_stem_level(data_stems, height_model)
calc_variables_tree_level(data_dendro, data_stems_calc)

}
