% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_variables_tree_level.R
\name{calc_variables_tree_level}
\alias{calc_variables_tree_level}
\title{Calculate additional variables on tree level}
\usage{
calc_variables_tree_level(data_dendro, data_stems_calc)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables \code{plot_id},
\code{plottype}, \code{tree_measure_id}, \code{date_dendro}, \code{dbh_mm}, \code{height_m},
\code{species}, \code{alive_dead}, \code{decaystage}, \code{period}, \code{old_id}, \code{year},
\code{subcircle}, \code{plotarea_ha},...
(output of function \code{load_data_dendrometry()})}

\item{data_stems_calc}{dataframe on stem level measurements with variables
\code{plot_id}, \code{plottype}, \code{tree_measure_id}, \code{date_dendro}, \code{dbh_mm},
\code{height_m}, \code{species}, \code{alive_dead}, \code{decaystage}, \code{period}, \code{year},
\code{subcircle}, \code{plotarea_ha},...
(output of function \code{calc_variables_stem_level()})}
}
\value{
Dataframe with ...
}
\description{
This function calculates additional variables based on measurements, such as
\itemize{
\item \code{nr_of_stems}: the number of shoots in the tree
(= 1 for an individual tree; >= 1 when coppice)
\item \code{individual}: true for individual tree or coppice,
false if record is a secondary shoot
\item \code{calc_height_m}: calculated height based on \code{dbh_mm} and
a species specific diameter-height model
\item \code{basal_area_m2}
\item \code{vol_bole_m3}: calculated based on \code{dbh_mm}, \code{calc_height_m} and
species specific tariffs
\item \code{vol_crown_m3}: calculated based on \code{dbh_mm} and
species specific tariffs
\item \code{vol_tot_m3}: sum of \code{vol_bole_m3} and \code{vol_crowwn_m3}
\item \code{dbh_mm} (based on average for coppice trees)
\item \code{decaystage} (based on average for coppice trees)
\item \code{basal_area_alive_m2_ha}
\item \code{basal_area_dead_m2_ha}
\item \code{vol_alive_m3_ha}
\item \code{vol_dead_standing_m3_ha}
\item \code{vol_bole_alive_m3_ha}
\item \code{vol_bole_dead_m3_ha}
}
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")

data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
# omit argument 'example_dataset = TRUE' below to use all height models
height_model <- load_height_models(example_dataset = TRUE)
data_stems_calc <- calc_variables_stem_level(data_stems, height_model)
calc_variables_tree_level(data_dendro, data_stems_calc)

}
