% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deadw_decay_plot_species.R
\name{calculate_deadw_decay_plot_species}
\alias{calculate_deadw_decay_plot_species}
\title{aggregate parameters by decay stage, plot, tree species and year}
\usage{
calculate_deadw_decay_plot_species(
  plotinfo,
  data_deadwood = NA,
  data_dendro_calc = NA
)
}
\arguments{
\item{plotinfo}{dataframe on surveyed plots with variables plot_id, plottype,
forest_reserve, survey_trees, survey_deadw, period and year_dendro
(output of function load_plotinfo())}

\item{data_deadwood}{dataframe on logs with variables plot_id, plottype,
date_dendro, species, decaystage, calc_volume_m3, period and year
(output of function load_data_deadwood())}

\item{data_dendro_calc}{dataframe on stems (shoots and trees) as given from
the function calc_variables_stem_level()}
}
\value{
dataframe with columns plot, year, tree_species, decaystage,
vol_log_m3_ha
}
\description{
This function calculates for each plot, tree species and year the volume
logs and standing dead wood per hectare and per decay stage.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")

data_deadwood <- load_data_deadwood(path_to_fieldmapdb)
data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models()
data_stems_calc <- calc_variables_stem_level(data_stems, height_model)
data_dendro_calc <- calc_variables_tree_level(data_dendro, data_stems_calc)
plotinfo <- load_plotinfo(path_to_fieldmapdb)
calculate_deadw_decay_plot_species(plotinfo, data_deadwood, data_dendro_calc)

}
