% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendro_plot.R
\name{calculate_dendro_plot}
\alias{calculate_dendro_plot}
\title{aggregate parameters by plot}
\usage{
calculate_dendro_plot(by_plot_year)
}
\arguments{
\item{by_plot_year}{dataframe with values for each plot and year, which is the result of the calculation by function calculate_dendro_plot_year()}
}
\value{
dataframe with columns plot, year_diff, number_of_tree_species_diff, number_of_trees_ha_diff, basal_area_m2_ha_diff, volume_stem_m3_ha_diff
}
\description{
This function compares for each plot the differences between years for: number of tree species, number of trees, basal area, and volume (calculated per hectare). It gives results for differences between subsequent measures (based on 'period') and between the last and the first measure.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_deadwood <-
  load_data_deadwood("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
by_plot_year <- calculate_dendro_plot_year(data_dendro, data_deadwood)
calculate_dendro_plot(by_plot_year)
}

}
