% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendro_plot.R
\name{calculate_dendro_plot}
\alias{calculate_dendro_plot}
\title{aggregate parameters by plot and year}
\usage{
calculate_dendro_plot(data_dendro_calc, data_deadwood, plotinfo)
}
\arguments{
\item{data_dendro_calc}{dataframe on tree measures with variables plot_id,
plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species,
alive_dead, decaystage, vol_tot_m3, basal_area_m2, period, OldID, year,
subcircle, plotarea_ha,... (output of function calc_variables_tree_level())}

\item{data_deadwood}{dataframe on logs with variables plot_id, plottype,
date_dendro, species, decaystage, calc_volume_m3, period and year
(output of function load_data_deadwood())}

\item{plotinfo}{dataframe on surveyed plots with variables plot_id, plottype,
forest_reserve, survey_trees, survey_deadw, period and year_dendro
(output of function load_plotinfo())}
}
\value{
dataframe with columns plot, year, number_of_tree_species,
number_of_trees_ha, basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot and year some values per hectare:
number of tree species, number of trees, basal area and volume.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
# (add path to your height models here)
path_to_height_models <-
  system.file("example/height_models", package = "forrescalc")

data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
data_deadwood <- load_data_deadwood(path_to_fieldmapdb)
data_stems <- compose_stem_data(data_dendro, data_shoots)
height_model <- load_height_models(path_to_height_models)
data_stems_calc <- calc_variables_stem_level(data_stems, height_model)
data_dendro_calc <- calc_variables_tree_level(data_dendro, data_stems_calc)
plotinfo <- load_plotinfo(path_to_fieldmapdb)
calculate_dendro_plot(data_dendro_calc, data_deadwood, plotinfo)

}
