% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendro_plot.R
\name{calculate_dendro_plot}
\alias{calculate_dendro_plot}
\title{aggregate parameters by plot and year}
\usage{
calculate_dendro_plot(data_dendro, data_deadwood)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, Plottype, tree_measure_id, date_dendro, DBH_mm, Height_m, species, AliveDead, decaystage, Vol_tot_m3, BasalArea_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}

\item{data_deadwood}{dataframe on logs with variables plot_id, Plottype, date_dendro, species, decaystage, CalcVolume_m3, period and year (output of function load_data_deadwood())}
}
\value{
dataframe with columns plot, year, number_of_tree_species, number_of_trees_ha, basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot and year some values per hectare: number of tree species, number of trees, basal area, and volume.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_deadwood <-
  load_data_deadwood("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_dendro_plot(data_dendro, data_deadwood)
}

}
