% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendrometry.R
\name{calculate_dendrometry}
\alias{calculate_dendrometry}
\title{aggregates parameters on dendrometry of trees}
\usage{
calculate_dendrometry(data_dendro, data_deadwood)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, Plottype, tree_measure_id, date_dendro, DBH_mm, Height_m, species, AliveDead, decaystage, Adjust_Vol_tot_m3, AdjustBasalArea_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}

\item{data_deadwood}{dataframe on logs with variables plot_id, Plottype, date_dendro, species, decaystage, CalcVolume_m3, period and year (output of function load_data_deadwood())}
}
\value{
List of dataframes that are mentioned in the above description
}
\description{
This function makes aggregations of individual tree measures on the levels of
\itemize{
 \item plot and year
 \item plot, tree species and year
 \item individual trees: statuses in different years
 \item diameter class, plot and year
 \item diameter class, plot, tree species and year
}
and it makes aggregations of volume data on logs on the levels of
\itemize{
 \item decay stage, plot and year
 \item decay stage, plot, tree species and year
}
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_deadwood <-
  load_data_deadwood("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_dendrometry(data_dendro, data_deadwood)
}

}
