% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendrometry.R
\name{calculate_dendrometry}
\alias{calculate_dendrometry}
\title{aggregates parameters on dendrometry of trees}
\usage{
calculate_dendrometry(data_dendro, data_deadwood)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, tree_measure_id, date_dendro, DBH_mm, Height_m, species, AliveDead, decaystage, Adjust_Vol_tot_m3, AdjustBasalArea_m2, period, OldID, year, plottype, plotarea_ha,... (output of function load_data_dendrometry())}

\item{data_deadwood}{dataframe on logs with variables plot_id, date_dendro, CalcVolume_m3, period and year (output of function load_data_deadwood())}
}
\value{
List of dataframes that are mentioned in the above description
}
\description{
This function makes aggregations of individual tree measures on the levels of
\itemize{
 \item plot and year
 \item plot, tree species and year
 \item individual trees: statuses in different years
}
}
