% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendrometry.R
\name{calculate_dendrometry}
\alias{calculate_dendrometry}
\title{Calcualtes and aggregates parameters on dendrometry of trees}
\usage{
calculate_dendrometry(data_dendro, data_deadwood, data_shoots, height_model)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species, alive_dead, decaystage, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}

\item{data_deadwood}{dataframe on logs with variables plot_id, plottype, date_dendro, species, decaystage, calc_volume_m3, period and year (output of function load_data_deadwood())}

\item{data_shoots}{dataframe on shoots as given from the function load_data_shoots()}

\item{height_model}{dataframe with `model` containing 'exp' or 'ln',
coeficients `P1` and `P2` to calculate height model for each combination of
`species`, `forest_reserve`, `period` and `plottype`. Height models in .xlsx
generated by Fieldmap can be grouped in a dataframe using function
`load_height_models()`}
}
\value{
List of dataframes that are mentioned in the above description
}
\description{
This function calculates additional variables and makes aggregations of individual tree measures on the levels of
\itemize{
 \item plot and year
 \item plot, tree species and year
 \item diameter class, plot and year
 \item diameter class, plot, tree species and year
}
and it makes aggregations of volume data on logs on the levels of
\itemize{
 \item decay stage, plot and year
 \item decay stage, plot, tree species and year
}
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_deadwood <-
  load_data_deadwood("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_shoots <-
  load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
height_model <- load_height_models("C:/bosreservaten/Hoogtemodellen/")
calculate_dendrometry(data_dendro, data_deadwood, data_shoots, height_model)
}

}
