% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dendrometry.R
\name{calculate_dendrometry}
\alias{calculate_dendrometry}
\title{Calculates and aggregates parameters on dendrometry of trees}
\usage{
calculate_dendrometry(
  data_dendro,
  data_deadwood,
  data_shoots,
  height_model,
  plotinfo
)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables \code{plot_id},
\code{plottype}, \code{tree_measure_id}, \code{date_dendro}, \code{dbh_mm}, \code{height_m},
\code{species}, \code{alive_dead}, \code{decaystage}, \code{period}, \code{old_id}, \code{year},
\code{subcircle}, \code{plotarea_ha},...
(output of function \code{load_data_dendrometry()})}

\item{data_deadwood}{dataframe on logs with variables \code{plot_id}, \code{plottype},
\code{date_dendro}, \code{species}, \code{decaystage}, \code{calc_volume_m3}, \code{period} and \code{year}
(output of function \code{load_data_deadwood()})}

\item{data_shoots}{dataframe on shoots as given from the function
\code{load_data_shoots()}}

\item{height_model}{dataframe with \code{model} containing 'exp' or 'ln',
coefficients \code{P1} and \code{P2} to calculate height model for each combination of
\code{species}, \code{forest_reserve}, \code{period} and \code{plottype}. Height models in .xlsx
generated by \code{Fieldmap} can be grouped in a dataframe using function
\code{load_height_models()}}

\item{plotinfo}{dataframe on surveyed plots with variables \code{plot_id},
\code{plottype}, \code{forest_reserve}, \code{survey_trees}, \code{survey_deadw}, \code{period} and
\code{year_dendro}
(output of function \code{load_plotinfo()})}
}
\value{
List of dataframes that are mentioned in the above description
}
\description{
This function calculates additional variables and makes aggregations of
individual tree measures on the levels of
\itemize{
\item plot and year
\item plot, tree species and year
\item diameter class, plot and year
\item diameter class, plot, tree species and year
}
and it makes aggregations of volume data on logs on the levels of
\itemize{
\item decay stage, plot and year
\item decay stage, plot, tree species and year
}
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")

data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_deadwood <- load_data_deadwood(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
# omit argument 'example_dataset = TRUE' below to use all height models
height_model <- load_height_models(example_dataset = TRUE)
plotinfo <- load_plotinfo(path_to_fieldmapdb)
calculate_dendrometry(
  data_dendro, data_deadwood, data_shoots, height_model, plotinfo)

}
