% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diam_plot_species_year.R
\name{calculate_diam_plot_species_year}
\alias{calculate_diam_plot_species_year}
\title{aggregate parameters by diameter class, plot, tree species and year}
\usage{
calculate_diam_plot_species_year(data_stems, data_deadwood)
}
\arguments{
\item{data_stems}{data on stems (shoots and trees) as given from the function compose_stem_data()}

\item{data_deadwood}{dataframe on logs with variables plot_id, date_dendro, species, decaystage, CalcVolume_m3, period and year (output of function load_data_deadwood())}
}
\value{
dataframe with columns plot, year, tree_species, DBHClass_5cm, basal_area_m2_ha, volume_m3_ha
}
\description{
This function calculates for each plot, tree species, year and diameter class some values per hectare: number of trees, basal area, and volume.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_shoots <-
  load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_stems <- compose_stem_data(data_dendro, data_shoots)
data_deadwood <-
  load_data_deadwood("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_diam_plot_species_year(data_stems, data_deadwood)
}

}
