% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration.R
\name{calculate_regeneration}
\alias{calculate_regeneration}
\title{aggregates parameters on regeneration of trees}
\usage{
calculate_regeneration(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables ...}
}
\value{
List of dataframes that are mentioned in the above description
}
\description{
This function makes aggregations of tree generation data on the levels of
\itemize{
 \item plot, height class and year
 \item plot, height class, tree species and year
 \item plot, height class and tree species (comparison between different years)
}
}
