% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_regeneration_core_area_height_species.R
\name{calculate_regeneration_core_area_height_species}
\alias{calculate_regeneration_core_area_height_species}
\title{aggregate regeneration parameters by plot, height, species and year}
\usage{
calculate_regeneration_core_area_height_species(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables ...}
}
\value{
dataframe with columns plot, species, year, height, min_number_of_trees_ha, max_number_of_trees_ha, number_of_subplots_with_regeneration and perc_subplots_with_regeneration
}
\description{
This function calculates for each plot, height, species and year the number of trees per ha, and the number and percentage of subplots in which the species is regenerating. This calculation is designed for core areas, that consist of different subplots.
}
\examples{
\dontrun{
#change path before running
data_regeneration_CA <-
  load_data_regeneration(
    "C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb",
    plottype = "Core area"
  )
calculate_regeneration_core_area_height_species(data_regeneration_CA)
}

}
