% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_regeneration_core_area_height_species.R
\name{calculate_regeneration_core_area_height_species}
\alias{calculate_regeneration_core_area_height_species}
\title{aggregate regeneration parameters by plot, height, species and year}
\usage{
calculate_regeneration_core_area_height_species(data_regeneration, plotinfo)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
plot_id, plottype, subplot_id, height_class, species, nr_of_regeneration,
rubbing_damage_number, period, year, subcircle, plotarea_ha,
min_number_of_regeneration and max_number_of_regeneration.}

\item{plotinfo}{dataframe on surveyed plots with variables plot_id, plottype,
forest_reserve, survey_trees, survey_deadw, period and year_dendro
(output of function load_plotinfo())}
}
\value{
dataframe with columns plot, species, year, height,
nr_of_subplots_with_regeneration, perc_subplots_with_regeneration,
rubbing_damage_perc, mean_number_of_regeneration_ha,
lci_number_of_regeneration_ha, uci_number_of_regeneration_ha and
approx_nr_regeneration_ha.
}
\description{
This function calculates for each plot, height, species and year the number
of regeneration per ha (or interval with mean and confidence interval
using a log transformation), and the number and percentage of
subplots in which the species is regenerating.
This calculation is designed for core areas, that consist of different
subplots.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_regeneration_CA <-
  load_data_regeneration(
    "C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb",
    plottype = "CA"
  )
plotinfo <-
  load_plotinfo("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_regeneration_core_area_height_species(data_regeneration_CA, plotinfo)
}

}
