% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_regeneration_core_area_height_species.R
\name{calculate_regeneration_core_area_height_species}
\alias{calculate_regeneration_core_area_height_species}
\title{aggregate regeneration parameters by plot, height, species and year}
\usage{
calculate_regeneration_core_area_height_species(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
plot_id, plottype, subplot_id, height_class, species, nr_of_regeneration,
rubbing_damage_number, period, year, subcircle, plotarea_ha,
min_number_of_regeneration and max_number_of_regeneration.}
}
\value{
dataframe with columns plot, species, year, height,
nr_of_subplots_with_regeneration, perc_subplots_with_regeneration,
rubbing_damage_perc, mean_number_of_regeneration_ha,
lci_number_of_regeneration_ha, uci_number_of_regeneration_ha and
approx_nr_regeneration_ha.
}
\description{
This function calculates for each plot, height, species and year the number
of regeneration per ha (or interval with mean and confidence interval
using a log transformation), and the number and percentage of
subplots in which the species is regenerating.
This calculation is designed for core areas, that consist of different
subplots.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration_CA <-
  load_data_regeneration(path_to_fieldmapdb, plottype = "CA")
calculate_regeneration_core_area_height_species(data_regeneration_CA)

}
