% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot.R
\name{calculate_regeneration_plot}
\alias{calculate_regeneration_plot}
\title{calculate species number by plot and year}
\usage{
calculate_regeneration_plot(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
plot_id, plottype, subplot_id, height_class, species, nr_of_regeneration,
rubbing_damage_number, period, year, subcircle, plotarea_ha,
min_number_of_regeneration and max_number_of_regeneration.}
}
\value{
dataframe with columns plot, subplot, year, period,
number_of_tree_species, nr_of_tree_species_established,
mean_number_established_ha, lci_number_established_ha,
uci_number_established_ha,
mean_number_seedlings_ha, lci_number_seedlings_ha,
uci_number_seedlings_ha,
mean_rubbing_damage_perc_established, lci_rubbing_damage_perc_established,
uci_rubbing_damage_perc_established,
mean_rubbing_damage_perc_seedlings, lci_rubbing_damage_perc_seedlings,
uci_rubbing_damage_perc_seedlings,
approx_nr_established_ha, approx_nr_seedlings_ha,
approx_rubbing_damage_perc_established, approx_rubbing_damage_perc_seedlings.
}
\description{
This function calculates for each plot and year the number of species, total
number of seedlings and established regeneration (or interval with mean
and confidence interval using a log transformation) and
rubbing damage percentage for seedlings and established regeneration.
For core area plots, these variables are calculated for each subplot.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration <- load_data_regeneration(path_to_fieldmapdb)
calculate_regeneration_plot(data_regeneration)

}
