% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot_height.R
\name{calculate_regeneration_plot_height}
\alias{calculate_regeneration_plot_height}
\title{calculate species number by plot, tree height class and year}
\usage{
calculate_regeneration_plot_height(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables ...}
}
\value{
dataframe with columns plot, year, height_class and number_of_tree_species
}
\description{
This function calculates for each plot, tree height class and year the number of species, total number of trees and game damage percentage for regeneration.
}
\examples{
\dontrun{
#change path before running
data_regeneration <-
  load_data_regeneration("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_regeneration_plot_height(data_regeneration)
}

}
