% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot_height.R
\name{calculate_regeneration_plot_height}
\alias{calculate_regeneration_plot_height}
\title{calculate species number by plot, tree height class and year}
\usage{
calculate_regeneration_plot_height(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
plot_id, plottype, subplot_id, height_class, species, nr_of_regeneration,
rubbing_damage_number, period, year, subcircle, plotarea_ha,
min_number_of_regeneration and max_number_of_regeneration.}
}
\value{
dataframe with columns plot, subplot, year, period, height_class,
number_of_tree_species, approx_rubbing_damage_perc,
mean_number_of_regeneration_ha,
lci_number_of_regeneration_ha, uci_number_of_regeneration_ha and
approx_nr_regeneration_ha.
}
\description{
This function calculates for each plot, tree height class and year
the number of species, total number of regeneration (or interval with mean
and confidence interval using a log transformation) and
approximate rubbing damage percentage for regeneration.
For core area plots, these variables are calculated for each subplot.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration <- load_data_regeneration(path_to_fieldmapdb)
calculate_regeneration_plot_height(data_regeneration)

}
