% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regeneration_plot_height_species.R
\name{calculate_regeneration_plot_height_species}
\alias{calculate_regeneration_plot_height_species}
\title{aggregate parameters by plot, tree height class, species and year}
\usage{
calculate_regeneration_plot_height_species(data_regeneration)
}
\arguments{
\item{data_regeneration}{dataframe on tree regeneration with variables
plot_id, plottype, subplot_id, height_class, species, nr_of_regeneration,
rubbing_damage_number, period, year, subcircle, plotarea_ha,
min_number_of_regeneration and max_number_of_regeneration.}
}
\value{
dataframe with columns plot, year, height_class, species,
rubbing_damage_perc, nr_of_regeneration_ha (in case exact numbers are
observed), mean_number_of_regeneration_ha (in case intervals are observed),
lci_number_of_regeneration_ha, uci_number_of_regeneration_ha and
approx_nr_regeneration_ha.
}
\description{
This function calculates for each plot, tree height class, species and year
the number of regeneration (or interval with mean and confidence interval)
and rubbing damage percentage per hectare for regeneration.
For core area plots, these variables are calculated for each subplot.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
data_regeneration <-
  load_data_regeneration("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_regeneration_plot_height_species(data_regeneration)
}

}
