% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vegetation.R
\name{calculate_vegetation}
\alias{calculate_vegetation}
\title{aggregates parameters on vegetation of forests}
\usage{
calculate_vegetation(data_vegetation, data_herblayer)
}
\arguments{
\item{data_vegetation}{dataframe on vegetation with variables ...}

\item{data_herblayer}{dataframe on vegetation in the species level ('herb layer') with variables ...}
}
\value{
List of dataframes that are mentioned in the above description
}
\description{
This function makes aggregations of vegetation data on the levels of
\itemize{
 \item plot and year
 \item subplot and year (only for plot type 'core area')
 \item plot, species and year (only for plot type 'core area')
}
}
\examples{
\dontrun{
#change path before running
data_vegetation <-
  load_data_vegetation("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_vegetation(data_vegetation)
}

}
