% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vegetation_plot_species.R
\name{calculate_vegetation_plot_species}
\alias{calculate_vegetation_plot_species}
\title{aggregate vegetation parameters by plot, species and year}
\usage{
calculate_vegetation_plot_species(data_herblayer)
}
\arguments{
\item{data_herblayer}{dataframe on vegetation in the species level ('herb layer') with variables ...}
}
\value{
dataframe with columns plot, species, year, number_of_subplots (= number of subplots where the species occurs), perc_of_subplots (= percentage of subplots with species), number_of_subplots_browsed, perc_of_subplots_browsed, number_of_subplots_seriously_browsed, perc_of_subplots_seriously_browsed and mean_coverage_class_average_perc
}
\description{
This function calculates for each plot, species and year the percentage of subplots in which the species is present and the percentage of subplots where the species is browsed (relative to the plots where it is present).  A difference is made between browsed (which contains all damage) and seriously browsed, which is reported if the damage is more than 1/20.  This calculation is designed for core areas, that consist of different subplots.
}
\examples{
\dontrun{
#change path before running
data_herblayer_CA <-
  load_data_herblayer(
    "C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb",
    plottype = "Core area"
  )
calculate_vegetation_plot_species(data_herblayer_CA)
}

}
