% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_stem_data.R
\name{compose_stem_data}
\alias{compose_stem_data}
\title{Combine dendro data and shoot data to give detailed stem data}
\usage{
compose_stem_data(data_dendro, data_shoots)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, Height_m, species, AliveDead, decaystage, Vol_tot_m3, BasalArea_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}

\item{data_shoots}{dataframe on shoots as given from the function load_data_shoots()}
}
\value{
Dataframe with shoot data
}
\description{
This function replaces in the given dendrometric data (result from function load_data_dendrometry())
the diameters from coppice trees by their separate stems given in the shoot data (result from function load_data_shoots()).
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_shoots <-
  load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
compose_stem_data(data_dendro, data_shoots)
}

}
