% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose stem_data.R
\name{compose_stem_data}
\alias{compose_stem_data}
\title{Combine dendro data and shoot data to give detailed stem data}
\usage{
compose_stem_data(data_dendro, data_shoots)
}
\arguments{
\item{data_dendro}{}
}
\value{
Dataframe with shoot data
}
\description{
This function replaces in the given dendrometric data (result from function load_data_dendrometry())
the diameters from coppice trees by their separate stems given in the shoot data (result from function load_data_shoots()).
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
data_shoots <-
  load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
compose_stem_data(data_dendro, data_shoots)
}

}
