% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_stem_data.R
\name{compose_stem_data}
\alias{compose_stem_data}
\title{Combine dendro data and shoot data to give detailed stem data}
\usage{
compose_stem_data(data_dendro, data_shoots, extra_variables = FALSE)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id,
plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species,
alive_dead, decaystage, period, OldID, year, subcircle, plotarea_ha,...
(output of function load_data_dendrometry())}

\item{data_shoots}{dataframe on shoots as given from the function
load_data_shoots()}

\item{extra_variables}{Should additional variables such as iufro_hght,
iufro_vital, iufro_socia, remark and common_remark be added?
Default is FALSE (no).
ATTENTION: some variables as iufro-classes and (common-)remark are
- for coppice - collected at shoot level.
To include these extra variables, it is necessary to indicate this argument in both
load-functions (load_data_dendrometry() and load_data_shoots()):
extra_variables = TRUE.}
}
\value{
Dataframe with shoot data
}
\description{
This function replaces in the given dendrometric data (result from function
load_data_dendrometry()) the diameters, height, decaystage and info on
intact/snag from coppice trees by their separate stems given in the shoot
data (result from function load_data_shoots()).
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("database/mdb_bosres.sqlite", package = "forrescalc")
data_dendro <- load_data_dendrometry(path_to_fieldmapdb)
data_shoots <- load_data_shoots(path_to_fieldmapdb)
compose_stem_data(data_dendro, data_shoots)

#to include iufro-classes and other additional variables:
data_dendro <-
  load_data_dendrometry(path_to_fieldmapdb, extra_variables = TRUE)
data_shoots <-
  load_data_shoots(path_to_fieldmapdb, extra_variables = TRUE)
compose_stem_data(data_dendro, data_shoots, extra_variables = TRUE)

}
