% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diam_statistics_forres_species.R
\name{create_diam_statistics_forres_species}
\alias{create_diam_statistics_forres_species}
\title{Give diameter statistics by forest reserve, species and year}
\usage{
create_diam_statistics_forres_species(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species, alive_dead, decaystage, vol_tot_m3, basal_area_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}
}
\value{
dataframe with columns forres, species, year and measures on diameter distribution (min, max, mean, median, Q1, Q3)
}
\description{
This function calculates the diameter distribution on the level of forest reserve, species and year some values per diameter class and hectare: number of trees, basal area, and volume.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
create_diam_statistics_forres_species(data_dendro)
}

}
