% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_overview_status.R
\name{create_overview_status}
\alias{create_overview_status}
\title{couple data of each individual tree in different years}
\usage{
create_overview_status(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species, alive_dead, decaystage, vol_tot_m3, basal_area_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}
}
\value{
a dataset with 1 record per measurement, containing the given data of each tree in different years
}
\description{
This function groups all given information on the life stages of an individual tree during different measures.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
create_overview_status(data_dendro)
}

}
