% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_access_to_git.R
\name{from_access_to_git}
\alias{from_access_to_git}
\title{copy table(s) from access db to git repository}
\usage{
from_access_to_git(database, tables, repo_path, push = FALSE, strict = TRUE)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{tables}{vector with table names of tables that should be moved}

\item{repo_path}{name and path of local git repository in which results
should be saved}

\item{push}{push commits directly to the remote on github?
Default is FALSE (no). (This option can only be used with SSH.)}

\item{strict}{keep default TRUE to update data without structural changes,
change to FALSE only if tables are structurally changed
(e.g. additional column, change in sorting order,...)}
}
\value{
No value is returned, the tables are saved in the git repository.
}
\description{
This function loads one or more tables from the access database
(or an SQLite database) and saves them in a git repository.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("database/mdb_bosres.sqlite", package = "forrescalc")
from_access_to_git(
  database = path_to_fieldmapdb,
  tables = c("qLayer", "qMossLondo"),
  repo_path = "C:/gitrepo/forresdat"
)

}
