% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_git_to_access.R
\name{from_git_to_access}
\alias{from_git_to_access}
\title{copy table(s) from git repository to access db}
\usage{
from_git_to_access(tables, repo_path, database, remove_tables = FALSE)
}
\arguments{
\item{tables}{vector with table names of tables that should be moved}

\item{repo_path}{name and path of local git repository in which results should be saved}

\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{remove_tables}{overwrite existing tables in database?  Default is
FALSE, which means tables are not overwritten/deleted unless this parameter
is explicitly put on TRUE.}
}
\value{
No value is returned, the tables are saved in the access database.
}
\description{
This function loads one or more tables from  a git repository and saves them
in an access database.
}
\examples{
\dontrun{
#change paths before running
library(forrescalc)
from_git_to_access(
  tables = c("qAliveDead", "qdecaystage"),
  repo_path = "C:/gitrepo/forresdat",
  database = "C:/db/testdb.accdb"
)
}

}
