% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_deadwood.R
\name{load_data_deadwood}
\alias{load_data_deadwood}
\title{retrieves data on logs from fieldmap database}
\usage{
load_data_deadwood(
  database,
  plottype = NA,
  forest_reserve = NA,
  extra_variables = FALSE,
  processed = TRUE
)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{plottype}{possibility to select only data for a certain plot type, e.g.
'CP' for Circular plot or 'CA' for Core area
(the default NA means that data from all plots are retrieved)}

\item{forest_reserve}{possibility to select only data for 1 forest reserve
by giving the name of the forest reserve (the default NA means that data
from all plots are retrieved)}

\item{extra_variables}{Should additional variables such as
remark and common_remark be added?
Default is FALSE (no).}

\item{processed}{Should only processed and surveyed data be added?
Defaults to TRUE (yes).}
}
\value{
Dataframe with data on logs
}
\description{
This function queries the given database to retrieve data on deadwood (logs)
(ready for use in `calculate_dendrometry()` function).
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
load_data_deadwood(path_to_fieldmapdb)

}
