% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_dendrometry.R
\name{load_data_dendrometry}
\alias{load_data_dendrometry}
\title{retrieve dendrometry data from fieldmap database}
\usage{
load_data_dendrometry(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap structure) including path}
}
\value{
Dataframe with dendrometry data
}
\description{
This function queries the given database to retrieve data on dendrometry (ready for use in calculate_dendrometry function).
}
\examples{
\dontrun{
#change path before running
load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
