% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_dendrometry.R
\name{load_data_dendrometry}
\alias{load_data_dendrometry}
\title{retrieve dendrometry data from fieldmap database}
\usage{
load_data_dendrometry(
  database,
  plottype = NA,
  forest_reserve = NA,
  extra_variables = FALSE
)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{plottype}{possibility to select only data for a certain plot type, e.g.
'CP' for Circular plot or 'CA' for Core area
(the default NA means that data from all plots are retrieved)}

\item{forest_reserve}{possibility to select only data for 1 forest reserve
by giving the name of the forest reserve (the default NA means that data
from all plots are retrieved)}

\item{extra_variables}{Should additional variables such as
x_m, y_m,
coppice_id, iufro_hght, iufro_vital, iufro_socia, remark and common_remark be added?
Default is FALSE (no).}
}
\value{
Dataframe with dendrometry data
}
\description{
This function queries the given database to retrieve data on dendrometry
(ready for use in calculate_dendrometry function).
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
