% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_herblayer.R
\name{load_data_herblayer}
\alias{load_data_herblayer}
\title{retrieve species specific vegetation data from \code{Fieldmap} database}
\usage{
load_data_herblayer(
  database,
  plottype = NA,
  forest_reserve = NA,
  processed = TRUE
)
}
\arguments{
\item{database}{name of \code{Fieldmap}/Access database (with specific \code{Fieldmap}
structure) including path}

\item{plottype}{possibility to select only data for a certain plot type, e.g.
'CP' for Circular plot or 'CA' for Core area
(the default NA means that data from all plots are retrieved)}

\item{forest_reserve}{possibility to select only data for 1 forest reserve
by giving the name of the forest reserve (the default NA means that data
from all plots are retrieved)}

\item{processed}{Should only processed and surveyed data be added?
Defaults to TRUE (yes).}
}
\value{
Dataframe with vegetation data on the species level ('herb layer'),
containing columns as species, coverage_id, browse_index_id, date_vegetation
(= date of survey of specific species, different for spring flora and other
flora in the same plot), year (= year of survey of specific species, possibly
different for spring flora and other flora), ....
}
\description{
This function queries the given database to retrieve data on vegetation
(ready for use in \code{calculate_vegetation()} function).
\code{year_main_survey} refers to year of the main vegetation survey
(source is table \code{vegetation}),
while \code{year} refers to year of recording of that specific species
(possibly different for spring flora; source is table \code{herblayer})
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
load_data_herblayer(path_to_fieldmapdb)

}
