% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_regeneration.R
\name{load_data_regeneration}
\alias{load_data_regeneration}
\title{retrieve regeneration data from fieldmap database}
\usage{
load_data_regeneration(
  database,
  plottype = NA,
  forest_reserve = NA,
  processed = TRUE
)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{plottype}{possibility to select only data for a certain plot type, e.g.
'CP' for Circular plot or 'CA' for Core area
(the default NA means that data from all plots are retrieved)}

\item{forest_reserve}{possibility to select only data for 1 forest reserve
by giving the name of the forest reserve (the default NA means that data
from all plots are retrieved)}

\item{processed}{Should only processed and surveyed data be added?
Defaults to TRUE (yes).}
}
\value{
Dataframe with regeneration data
}
\description{
This function queries the given database to retrieve data on regeneration
(ready for use in calculate_regeneration function).
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
load_data_regeneration("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
load_data_regeneration(
  "C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb",
  plottype = "CP"
)
}

}
