% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_shoots.R
\name{load_data_shoots}
\alias{load_data_shoots}
\title{retrieve data on shoots from fieldmap database}
\usage{
load_data_shoots(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap structure) including path}
}
\value{
Dataframe with shoot data
}
\description{
This function queries the given database to retrieve additional data on shoots to use with dendrometry data (for functions on calculations of diameters).
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
load_data_shoots("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
