% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_shoots.R
\name{load_data_shoots}
\alias{load_data_shoots}
\title{retrieve data on shoots from fieldmap database}
\usage{
load_data_shoots(database, extra_variables = FALSE)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{extra_variables}{Should additional variables such as iufro_hght,
iufro_vital, iufro_socia, remark and common_remark be added?
Default is FALSE (no).}
}
\value{
Dataframe with shoot data
}
\description{
This function queries the given database to retrieve additional data on
shoots to use with dendrometry data.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
load_data_shoots(path_to_fieldmapdb)

}
