% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_vegetation.R
\name{load_data_vegetation}
\alias{load_data_vegetation}
\title{retrieve vegetation data from fieldmap database}
\usage{
load_data_vegetation(database, plottype = NA, forest_reserve = NA)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{plottype}{possibility to select only data for a certain plot type, e.g. 'Circular plot'
or 'Core area' (the default NA means that data from all plots are retrieved)}

\item{forest_reserve}{possibility to select only data for 1 forest reserve
by giving the name of the forest reserve (the default NA means that data
from all plots are retrieved)}
}
\value{
Dataframe with vegetation data, containing columns as total_herb_cover, total_shrub_cover, total_tree_cover, total_soildisturbance_game, date_vegetation (= date of vegetation survey), year_record (= year of vegetation survey), year (= year of vegetation survey, derived from date_vegetation if available, otherwise from year_record), ....
}
\description{
This function queries the given database to retrieve data on vegetation (ready for use in calculate_vegetation function).
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
load_data_vegetation("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
}

}
