% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_vegetation.R
\name{load_data_vegetation}
\alias{load_data_vegetation}
\title{retrieve vegetation data from fieldmap database}
\usage{
load_data_vegetation(
  database,
  plottype = NA,
  forest_reserve = NA,
  processed = TRUE
)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}

\item{plottype}{possibility to select only data for a certain plot type, e.g.
'CP' for Circular plot or 'CA' for Core area
(the default NA means that data from all plots are retrieved)}

\item{forest_reserve}{possibility to select only data for 1 forest reserve
by giving the name of the forest reserve (the default NA means that data
from all plots are retrieved)}

\item{processed}{Should only processed and surveyed data be added?
Defaults to TRUE (yes).}
}
\value{
Dataframe with vegetation data, containing columns as
`total_herb_cover`, `total_shrub_cover`, `total_tree_cover`,
`total_soildisturbance_game`,
`date_vegetation` (= date of vegetation survey),
`year_main_survey` (= year of vegetation survey), ....
}
\description{
This function queries the given database to retrieve data on vegetation
(ready for use in calculate_vegetation function).
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
load_data_vegetation(path_to_fieldmapdb)

}
