% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_vegetation.R
\name{load_data_vegetation}
\alias{load_data_vegetation}
\title{retrieves vegetation data from fieldmap database}
\usage{
load_data_vegetation(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap structure) including path}
}
\value{
Dataframe with regeneration data
}
\description{
This function queries the given database to retrieve data on vegetation (ready for use in calculate_vegetation function).
}
