% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_height_models.R
\name{load_height_models}
\alias{load_height_models}
\title{retrieve height model data from git repository forresheights}
\usage{
load_height_models()
}
\value{
Dataframe with height model data
}
\description{
This function groups the information on height models from the `.csv` files
in the git repository [forresheights](https://github.com/inbo/forresheights)
together in one dataframe.
}
\examples{
library(forrescalc)
load_height_models()

}
