% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_plotinfo.R
\name{load_plotinfo}
\alias{load_plotinfo}
\title{retrieve generic plot data from fieldmap database}
\usage{
load_plotinfo(database)
}
\arguments{
\item{database}{name of fieldmap/access database (with specific fieldmap
structure) including path}
}
\value{
Dataframe with columns plot_id, plottype, forest_reserve, period,
year of dendrometric survey and information on
(1) whether there has been a dendro, deadwood, regeneration and/or
vegetation survey and (2) whether the data have been processed or not.
}
\description{
This function queries the given database to retrieve additional data on plots
to save in forresdat and link with the datasets that are saved there.
}
\examples{
library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
load_plotinfo(path_to_fieldmapdb)

}
