% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forresdat.R
\name{read_forresdat}
\alias{read_forresdat}
\title{load datapackage from git repository forresdat}
\usage{
read_forresdat()
}
\value{
A frictionless datapackage with all tables and metadata from GitHub
repository forresdat, which can be explored using package
[frictionless](https://docs.ropensci.org/frictionless/).
To be able to recall the version of the data, this datapackage contains
an attribute with the version number of the release of forresdat from which
the data are taken.
}
\description{
This function reads the datapackage from git repository forresdat
(and saves the forresdat data to a local temp directory to avoid unneeded
downloading in the future).
This datapackage contains both data and metadata and can be explored using
functions of the [frictionless](https://docs.ropensci.org/frictionless/)
package.

Data available in forresdat only contain observations, so no records with
zero values are added for for instance species that were not observed and
hence absent.
These zero value records can easily be added by using the function
`add_zeros()`.

The different tables of this dataset contain data that are collected
using 2 different methods (plottypes):
circular plots (CP) and core areas (CA).
It is advised to only use one of them for analyses, as the data are likely
to differ due to method related differences.

General information on the plot level is available in table `plotinfo`,
which can easily be joined to other tables on `plot_id` and `period`
(or only `plot_id` if `period` is absent).
}
\examples{
library(forrescalc)
datapackage <- read_forresdat()
frictionless::resources(datapackage)
attr(datapackage, "forresdat")

}
