% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forresdat.R
\name{read_forresdat}
\alias{read_forresdat}
\title{load tables from git repository forresdat}
\usage{
read_forresdat(tablename, repo_path, join_plotinfo = TRUE)
}
\arguments{
\item{tablename}{name of the table that should be read}

\item{repo_path}{name and path of local git repository from which data should be retrieved}

\item{join_plotinfo}{should table plotinfo be joined to the chosen table to
add columns plottype and forest_reserve?  Default is TRUE.  (This is only
possible if the given table contains a column plot_id, so this parameter
should be put FALSE if this column is absent.)}
}
\value{
A dataframe with the specified table, default columns plottype and forest_reserve.
}
\description{
This function reads a table in git2rdata format from git repository forresdat
(or another git repository dependent on the given repo path).
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
read_forresdat(tablename = "dendro_by_plot", repo_path = "C:/gitrepo/forresdat")
}

}
