% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forresdat.R
\name{read_forresdat}
\alias{read_forresdat}
\title{load tables from git repository forresdat}
\usage{
read_forresdat(
  tablename,
  join_plotinfo = TRUE,
  plottype = c("CP", "CA", "all")
)
}
\arguments{
\item{tablename}{name of the table that should be read}

\item{join_plotinfo}{should table plotinfo be joined to the chosen table to
add columns forest_reserve, survey_dendro/deadw/reg/veg (TRUE or
FALSE) and data_processed (TRUE or FALSE)?
Default is TRUE.
(This is only possible if the given table contains a column plot_id,
so this parameter should be put FALSE if this column is absent.)
Must be FALSE if you want to load the table "plotinfo" itself.}

\item{plottype}{Data of which 'plottype' (used method) should be retrieved?
Default is 'CP' or 'circle plot', alternatively 'CA' or 'core area', or 'all'
(retrieve both circle plots and core areas) could be chosen.}
}
\value{
A dataframe with the specified table, default columns plottype,
forest_reserve, survey_dendro/deadw/reg/veg (TRUE or FALSE) and
data_processed (TRUE or FALSE)
}
\description{
This function reads a table in tsv format from git repository forresdat.
Data available in forresdat only contains observations, so no records with
zero values are added for for instance species that were not observed and
hence absent.
These zero value records can easily be added by using the function
`add_zeros()`.
To load table plotinfo, set argument `join_plotinfo = FALSE`.
}
\examples{
library(forrescalc)
read_forresdat(tablename = "dendro_by_plot")

}
