% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forresdat.R
\name{read_forresdat}
\alias{read_forresdat}
\title{load tables from git repository forresdat}
\usage{
read_forresdat(
  tablename,
  repo_path,
  join_plotinfo = TRUE,
  plottype = c("CP", "CA")
)
}
\arguments{
\item{tablename}{name of the table that should be read}

\item{repo_path}{name and path of local git repository from which data
should be retrieved}

\item{join_plotinfo}{should table plotinfo be joined to the chosen table to
add columns plottype, forest_reserve, survey_dendro/deadw/reg/veg (TRUE or
FALSE) and data_processed (TRUE or FALSE)?  Default is TRUE.  (This is only
possible if the given table contains a column plot_id, so this parameter
should be put FALSE if this column is absent.)}

\item{plottype}{Data of which 'plottype' (used method) should be retrieved?
Default is 'CP' or 'circle plot', alternatively 'CA' or 'core area' could be
chosen.}
}
\value{
A dataframe with the specified table, default columns plottype,
forest_reserve, survey_dendro/deadw/reg/veg (TRUE or FALSE) and
data_processed (TRUE or FALSE)
}
\description{
This function reads a table in git2rdata format from git repository forresdat
(or another git repository dependent on the given repo path).
Data available in forresdat only contains observations, so no records with
zero values are added for for instance species that were not observed and
hence absent.
These zero value records can easily be added by using the function `add_zeros()`.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
read_forresdat(tablename = "dendro_by_plot",
repo_path = "C:/gitrepo/forresdat")
}

}
