% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forresdat_table.R
\name{read_forresdat_table}
\alias{read_forresdat_table}
\title{load tables from git repository forresdat}
\usage{
read_forresdat_table(
  tablename,
  join_plotinfo = TRUE,
  plottype = c("CP", "CA", "all")
)
}
\arguments{
\item{tablename}{name of the table that should be read}

\item{join_plotinfo}{should table plotinfo be joined to the chosen table to
add columns forest_reserve, survey_dendro/deadw/reg/veg (TRUE or
FALSE) and data_processed (TRUE or FALSE)?
Default is TRUE.
(This is only possible if the given table contains a column plot_id,
so this parameter should be put FALSE if this column is absent.)
Must be FALSE if you want to load the table "plotinfo" itself.}

\item{plottype}{Data of which 'plottype' (used method) should be retrieved?
Default is 'CP' or 'circle plot', alternatively 'CA' or 'core area', or 'all'
(retrieve both circle plots and core areas) could be chosen.}
}
\value{
A dataframe with the specified table, default columns plottype,
forest_reserve, survey_dendro/deadw/reg/veg (TRUE or FALSE) and
data_processed (TRUE or FALSE).
To be able to recall the version of the data, this dataframe contains
an attribute with the version number of the release of forresdat from which
the data are taken.
}
\description{
This function reads a table in csv format from git repository forresdat
(and saves the forresdat data to a local temp directory to avoid unneeded
downloading in the future).
Data available in forresdat only contains observations, so no records with
zero values are added for for instance species that were not observed and
hence absent.
These zero value records can easily be added by using the function
`add_zeros()`.
To load table plotinfo, set argument `join_plotinfo = FALSE`.
}
\examples{
library(forrescalc)
data_dendro <- read_forresdat_table(tablename = "dendro_by_plot")
data_dendro
attr(data_dendro, "forresdat")

}
