% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results_access.R
\name{save_results_access}
\alias{save_results_access}
\title{save results of calculations in Access database}
\usage{
save_results_access(results, database, remove_tables = FALSE)
}
\arguments{
\item{results}{results from calculations in package forrescalc as a named list}

\item{database}{name of (empty) Access database including path in which
results should be saved}

\item{remove_tables}{overwrite existing tables in database?  Default is
FALSE, which means tables are not overwritten/deleted unless this parameter
is explicitly put on TRUE.}
}
\value{
No value is returned, data are saved in the specified database
}
\description{
This function saves the results from calculations in the forrescalc package
(or any other named list with dataframes) in an Access database.
List item names will be used to name each of the tables,
which contain as a content the different dataframes.
}
\examples{
\dontrun{
#change path before running
library(forrescalc)
database <- system.file("database/mdb_bosres.sqlite", package = "forrescalc")
data_dendro <- load_data_dendrometry(database)
data_deadwood <- load_data_deadwood(database)
data_shoots <- load_data_shoots(database)
height_model <- load_height_models("C:/bosreservaten/Hoogtemodellen/")
result_dendro <-
  calculate_dendrometry(data_dendro, data_deadwood, data_shoots, height_model)
save_results_access(result = result_dendro, database = "C:/db/testdb.accdb")
#Repeating the previous line of code will give an error, because you try to
#overwrite a table that was already saved in the database on the first run.
#To overwrite previously saved tables, use this command:
save_results_access(result = result_dendro, database = "C:/db/testdb.accdb", remove_tables = TRUE)
}

}
