% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results_git.R
\name{save_results_git}
\alias{save_results_git}
\title{save results of calculations in git repository}
\usage{
save_results_git(results, repo_path)
}
\arguments{
\item{results}{results from calculations in package forrescalc as a named list}

\item{repo_path}{name and path of local git repository in which results should be saved}
}
\value{
No value is returned, data are saved in the specified git repository
}
\description{
This function saves the results from calculations in the forrescalc package (or any other named list with dataframes) in a git repository.  List item names will be used to name each of the tables, which contain as a content the different dataframes.
}
