% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_status.R
\name{summarise_status}
\alias{summarise_status}
\title{give the statuses of each individual in different years}
\usage{
summarise_status(data_dendro)
}
\arguments{
\item{data_dendro}{dataframe on tree measures with variables plot_id, Plottype, tree_measure_id, date_dendro, DBH_mm, Height_m, species, AliveDead, decaystage, Adjust_Vol_tot_m3, AdjustBasalArea_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())}
}
\value{
a dataset with 1 record for each tree, containing the decay stage in different years
}
\description{
This function groups all information on the life stages of an individual tree during different measures.
}
\examples{
\dontrun{
#change path before running
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
summarise_status(data_dendro)
}

}
